requirements.txt
streamlit
pandas
sqlite3
streamlit_authenticator
bcrypt
geopy.geocoders
folium
streamlit_folium
random